+++ back/src/main/java/com/baba/back/baby/invitation/repository/InvitationCodeRepository.java
@@ -1,6 +1,6 @@
-package com.baba.back.invitation.repository;
+package com.baba.back.baby.invitation.repository;
 
-import com.baba.back.invitation.domain.InvitationCode;
+import com.baba.back.baby.invitation.domain.InvitationCode;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface InvitationCodeRepository extends JpaRepository<InvitationCode, Long> {
